IOS: Rename MemoryValues to VersionInfo
