+++ Source/Core/Core/IOS/VersionInfo.cpp
@@ -2,7 +2,7 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
-#include "Core/IOS/MemoryValues.h"
+#include "Core/IOS/VersionInfo.h"
 
 #include <array>
