+++ Source/Core/Core/IOS/VersionInfo.h
